Refactored logs in nexus_structure
